pyproject.toml
[project]
name = "hue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "hue-entertainment-pykit>=0.9.3",
    "jinja2>=3.1.6",
    "numpy>=2.3.5",
    "pyaudio>=0.2.14",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "python-dotenv>=1.2.1",
    "python-mbedtls==2.10.1",
    "requests>=2.31.0",
    "uvicorn[standard]>=0.42.0",
    "websockets>=16.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hue_visualizer"]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.5",
]
